requirements.txt
aiodns==3.0.0
aiohttp==3.8.1
aiosignal==1.2.0
appdirs==1.4.4
asgiref==3.5.2
async-generator==1.10
async-timeout==4.0.2
attrs==22.1.0
beautifulsoup4==4.11.1
bs4==0.0.1
ccxt==1.93.1
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.1
country-currencies==0.2
cryptography==37.0.4
decorator==5.1.1
Django==4.0
django-extensions==3.2.0
django-filter==22.1
django-rest-framework==0.1.0
djangorestframework==3.13.1
dnspython==2.0.0
forex-python==1.8
frozenlist==1.3.1
greenlet==1.1.3
h11==0.13.0
idna==3.3
inflection==0.5.1
investpy==1.0.7
ipwhois==1.2.0
lxml==4.9.1
more-itertools==8.14.0
multidict==6.0.2
multitasking==0.0.11
numpy==1.23.2
outcome==1.2.0
pandas==1.4.4
pandas-datareader==0.10.0
psycopg2-binary==2.9.3
py4j==0.10.9.5
pycares==4.2.2
pycoingecko==3.0.0
pycountry==22.3.5
pycparser==2.21
PySocks==1.7.1
pyspark==3.3.0
python-dateutil==2.8.2
pytz==2022.2.1
PyYAML==6.0
Quandl==3.7.0
requests==2.28.1
selenium==4.4.3
simplejson==3.17.6
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
SQLAlchemy==1.4.40
sqlparse==0.4.2
tabulate==0.8.10
trio==0.21.0
trio-websocket==0.9.2
Unidecode==1.3.4
urllib3==1.26.12
wbdata==0.3.0
wsproto==1.2.0
yarl==1.7.2
yfinance==0.1.74
